### 2025-08-04 #8709 @editorialbot
[PRE REVIEW]: exa-AMD: A Scalable Workflow for Accelerating AI-Assisted Materials Discovery and Design
**Submitting author:** @MoraruMaxim (<a href="http://orcid.org/0000-0001-7213-089X">Maxim Moraru</a>)
**Repository:** https://github.com/ML-AMD/exa-amd
**Branch with paper.md** (empty if default branch):
**Version:** v0.1
