### 2025-08-26 #8709 @espottesmith
3 weeks is totally reasonable, as far as I'm concerned.

@MoraruMaxim, is ~3 weeks an acceptable time period for you?
