### 2025-08-27 #8709 @ChrisWoodgate
Hi @espottesmith!

Thanks for considering me for this review.

Unfortunately, I am out of the office for the next three weeks with various travel/conference/workshop commitments, so wouldn't be able to make a start on the review until mid- to late-September.

> Three weeks is fine. Just to confirm, [@PatrickJTaylor](https://github.com/PatrickJTaylor) do you currently have a VASP license? (Evaluating the full capabilities of the workflow requires one.)

Additionally, I don't have a `VASP` license. (I usually use `CASTEP` for my plane wave pseudopotential calculations as it has a free worldwide academic license! 😁)

So, I think it is probably best that I sit this one out—sorry!

It might be worth talking to @jameskermode, as this seems right up his street. I am fairly sure that there is a `VASP` license floating around at Warwick and he might have a postdoc/senior PhD student willing to review.
