### 2025-09-05 #8879 @alchem0x2A
@espottesmith I've provided my comments to the authors as issues created in the source code repo
