### 2025-09-15 ML-AMD/exa-amd#25 @PatrickJTaylor
[JOSS Review] Questions and suggestions for the paper
Linking openjournals/joss-reviews#8879 as this is part of the ongoing JOSS review.

Having read the current draft of the JOSS paper, I have a question for the authors, as well as some suggested changes.

My question echoes what has already been highlighted in #20, namely: is `exa-AMD` aiming to offer an alternative to other automation frameworks in the materials modelling space such as `atomate2`, `AiiDA` etc, or is it intended as a pre-configured materials discovery workflow? I feel that the current draft of the paper straddles the line between these two possibilities, and this makes the primary purpose of `exa-AMD` somewhat unclear to the reader. Most of my suggestions stem from this point.

# Suggestions
1. Add a new section, e.g., `Related software`, and use this to explain where `exa-AMD` fits in with other workflow automation tools relevant to materials modelling. Such a section could be used to answer the question I posed above, and thus clarify the exact purpose of `exa-AMD`.
2. Elaborate upon Figure 1, in particular: clarify the novelty of the predicted Ce-Fe-In phase diagram and the specific relevance of Ce3FeIn - is this predicted to be thermodynamically stable? If this figure is supposed to represent the novel results of a research project that was enabled by `exa-AMD`, then I would suggest that the figure caption be significantly expanded for clarity and that this research should be briefly highlighted in the main text.
3. Similar in spirit to 1, the caption of Figure 2 could be further clarified: is this *the only* workflow that `exa-AMD` is intended to provide, or is it just a pre-configured example that is intended to be built upon?

I would also like to echo the changes already suggested in #20, many of which mirror my own thoughts.
