### 2025-09-15 ML-AMD/exa-amd#26 @PatrickJTaylor
[JOSS Review] Scope of tests
Linking openjournals/joss-reviews#8879 as this is part of the ongoing JOSS review.

As far as I can tell, only a small number of unit tests have been written for this codebase. It is good to see that these tests are configured to run as part of continuous integration, but I am somewhat concerned that much of the core functionality of the software remains untested.

Could the authors please clarify: approximately how much of the total codebase is unit tested? It looks to me like a relatively low proportion, but please correct me if I am missing some tests or am misunderstanding how the current tests have been implemented.

Regardless of the above, I would suggest:
1. Collect code coverage as part of CI and (optionally) publish this to the README with a Markdown badge or similar. Remedying #21 will likely make this significantly easier.
2. If test coverage is indeed quite low, expand the test suite to include many more functions.
3. Include more tests that verify the *quantitative* predictions of `exa-AMD` workflows. For example, the `test_cgcnn` function appears to validate the CGCNN model simply by ensuring that the output CSV has the correct number/arrangement of entries, but does not seem to actually verify that the model's predictions are numerically consistent between test runs.
