### 2025-09-15 ML-AMD/exa-amd#27 @PatrickJTaylor
[JOSS Review] API documentation appears to be incomplete
Linking openjournals/joss-reviews#8879 as this is part of the ongoing JOSS review.

The API-level documentation for this software appears to be incomplete in the following sense:
1. The documentation currently available at [https://ml-amd.github.io/exa-amd/api.html](https://ml-amd.github.io/exa-amd/api.html) is sometimes very minimal e.g., arguments are not always explained and sometimes appear to be mistyped, summaries of functionality can be overly concise, there is a lack of example inputs etc.
2. Many functions present in the codebase are not represented in the API documentation at all. I suspect some of this is a consequence of #21 and that proper packaging would make auto-generating most of the API docs significantly easier. To be clear, I do not think that 100% documentation of the API is strictly necessary, but at the moment there are some unusual omissions (e.g., the code in `cms_dir` is not in the docs, despite being highly relevant to the core functionality of the software).
