### 2025-09-15 ML-AMD/exa-amd#28 @PatrickJTaylor
[JOSS Review] Contributing guidelines
Linking openjournals/joss-reviews#8879 as this is part of the ongoing JOSS review.

It is good to see that the README for this project specifies that contributors may wish to open pull requests, but it would be nice to see a more detailed set of contributing guidelines.

It may be helpful to include the following in a `CONTRIBUTING.md` or similar:
1. Guidelines as to code style and organisation (e.g., adherence to PEP8, docstring style etc)
2. Checks to potentially run before submitting a PR (linting, testing, type checking etc)
3. The types of changes that may be particularly welcome (e.g., new workflows, interfaces to new models)
4. A recommended "developer workflow" to follow pre PR submission
