### 2025-09-16 ML-AMD/exa-amd#29 @yw-fang
[REVIEW comments] for the paper and code submitted to JOSS
This submitted work to the JOSS (https://github.com/openjournals/joss-reviews/issues/8879) is a very timely work and code-implementation based on the state of the art of computational tools, which is expected to accerelate the computational materails discovery.

Herein, I list several comments on the manuscript and code itself that you may consider addressing before the accept for publication.

- [x] It is very user-friendly that the authors have provided a YAML file (amd_env.yml) for creating the virtual environment. However, it may be clearer to state explicitly that this file is included in the source code repository, rather than referring to it as "... environment yaml file we provide" in the README.md file.

- [x] To broaden the usability of the code, it could be beneficial to interface it with additional ab initio packages for the ab initio calculations beyond the current VASP implementation. Do the authors have plans to extend the interface? If not, could they provide contribution guidelines so that researchers in the community can more easily contribute? Because this current code exa-AMD is an open source code, I think it would be more powerful it can be interfaced to some open source code such as Quantum Espresso in addition to the property software VASP.

- [x] In the section of **Table of Contents** in README.md, the "Prerequisites" and "Usage" are clickable and brong me to the subsection as expected. But the other two ones do not work when clicking. I tested both Chrome and Firefox in linux systems.

- [x] This is only a minor comment that the authors may optionally consider addressing. If I am not mistaken, the current implementation only supports calculations at 0 GPa. Will it also be able to support calculations under pressure?

- [x] It is noticed that there are more functions defined in the source code that the ones shown in the API webpage: https://ml-amd.github.io/exa-amd/api.html  Did the authors particularly selected some ones?

- [x] Because this code also makes use of external packages such as CGCNN, it raises the question of how to properly cite it when using the code in research. Could the authors add a section in the README.md page explaining how to cite the code? (Of course, this JOSS paper to be published in the near future will be cited, but if there are other tools that should also be acknowledged, it would be helpful to provide citation instructions.)

- [x] In README.md, the present link of "[Parsl's official documentation](https://parsl.readthedocs.io/en/stable/userguide/configuration/execution.html)" does not work and returns "404 Documentation page not found"

## Manuscript

- [x] I noticed there are multiple authors including highly senior authors. However, according to the policy of te authorship of JOSS, "Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions". To ensure transparency, it would be helpful to clearly indicate the specific contributions of each co-author, either in the manuscript itself or on the project’s code homepage.

- [x] Figure 1 was not cited and discussed in the manuscript. Please provide brief discussions

- [x] In high-throughput crystal structure prediction, it is important to remove duplicate structures. The authors indicate that this was done at the filtering stage based on structural similarity. Could the authors expand on this point with more details? In addition, if any existing techniques were used to identify identical crystal structures, appropriate references should be provided.

- [x] This manuscript does not indicate whether there are any codes similar to exa-AMD. Could the authors provide an introduction to the current state of the art? If similar codes already exist, please clarify what makes this code unique and why it merits publication in JOSS.
