### 2025-09-16 #8879 @yw-fang
I would like to thank the editor and the authors for your patience. My first review report is done at https://github.com/ML-AMD/exa-amd/issues/29
@espottesmith
