### 2025-09-18 ML-AMD/exa-amd#30 @PatrickJTaylor
[JOSS Review] Suggested improvements to high-level documentation
Linking openjournals/joss-reviews#8879 as this is part of the ongoing JOSS review.

Having now played around with the software using a couple of different HPC stacks, I have some suggested improvements for the `exa-AMD` documentation.

1. Whilst the general workflow that `exa-AMD` provides out-of-the-box is nicely summarised in the JOSS manuscript, I think that further details would be helpful. For example, how exactly are structures de-duplicated/what is the metric used to judge structural similarity? According to what standards have the DFT inputs been chosen for all VASP calculations (choice of xc-functional, size of plane wave basis, k-point density etc)? I don't think that all such details need to be in the JOSS manuscript, but I do think that they need to be publicly available as part of the `exa-AMD` documentation in some form.
2. The tutorial in the documentation is a nice starting point, but ultimately I think it assumes a lot of prior knowledge on the part of the reader. Despite having used similar materials modelling workflow automation software before, I had never used or configured `parsl` before trying out `exa-AMD`. This made my experience quite tedious at times, because I found myself regularly having to consult the `parsl` documentation for specific configuration details: the `exa-AMD` documentation was usually not sufficient in an of itself. Of course, I accept that this is unavoidable to some extent, it is very difficult to write workflow software that is truly HPC-agnostic, but nonetheless I feel that more could be done to make configuring `exa-AMD` easier on a wide variety of systems e.g, provide some detailed examples of how one might configure `exa-AMD` for a given HPC. The included configurations for `perlmutter` and such are helpful, but I think that various details could be better explained as part of the documentation. For example, a brief overview of `parsl`'s layers of abstraction would be helpful for those new to it: workers/blocks/providers etc. I know that the `parsl` documentation already exists for these things, but if you want `exa-AMD` to be generally user-friendly to the wider community, then a brief summary of the essentials would be helpful.
