### 2025-09-18 #8879 @PatrickJTaylor
@espottesmith I have now finished my initial review - thank you for your patience!

My comments and suggestions have been made as issues over in the `exa-AMD` repo, but I will also briefly summarise my main thoughts here for the benefit of @morarumaxim and the other authors.

I think that `exa-AMD` has clear research applications and I enjoyed my time testing it out on a few different HPC systems. That being said, I do think that various improvements ought to be made to the codebase/documentation/paper before publication in JOSS. In particular, I think the following issues deserve particular attention (especially point 1, which has implications for how we ought to approach the software as reviewers in the first place):

1. The **purpose/target audience for `exa-AMD` should be clarified**, with both the JOSS manuscript and the general documentation being adapted accordingly. It is clear that the software provides a materials discovery workflow with useful research applications, but it is not obvious if it is intended to go beyond this one pre-configured workflow. Parts of the JOSS manuscript seem to imply that `exa-AMD` is quite flexible and can easily be modified to introduce new workflows, perhaps in a manner similar to something like `AiiDA` and its plugins. From my experience of the software, its source code and the documentation, I am not sure that `exa-AMD` offers much to users that would make it especially simple to extend to new workflows/codes beyond what `parsl` gives you by default - I would be curious as to the authors' thoughts on this.
2. The **test suite and API-level documentation should both be expanded**, particularly with respect to core functionality that is currently absent. The most obvious example is the code in `cms_dir`, which is key to the primary `exa-AMD` workflow, but is not unit tested or present in the API docs.
3. **The documentation should include more detailed examples and explanations as to how to configure `exa-AMD` for different HPC systems**. The included configurations for `perlmutter` etc are a good start, but a lot of knowledge is currently assumed on the part of the user, particularly in terms of how `parsl` works.
4. **Proper guidelines for potential contributors should be made available via a `CONTRIBUTING.md` or similar**. The README already invites pull requests, which is good to see, but it is unclear where I should start as a developer if I wanted to make a substantial contribution to the codebase.

I should add that I agree with all of @alchem0x2A and @yw-fang's various comments too.
