### 2025-10-07 #8879 @MoraruMaxim
Hello,
We’ve completed the code refactor and made our first release!

Our current focus is on improving the documentation. I’m sorry this process has taken longer than expected. I’m aiming to have most of the remaining issues resolved by the end of the week.
