### 2025-10-07 ML-AMD/exa-amd#26 @MoraruMaxim
Here is our code coverage as part of CI:
```
Name                                Stmts   Miss  Cover
-------------------------------------------------------
ml_models/__init__.py                   0      0   100%
ml_models/cgcnn/__init__.py             2      0   100%
ml_models/cgcnn/data.py               134     37    72%
ml_models/cgcnn/model.py               64      9    86%
ml_models/cgcnn/predict.py            180     53    71%
parsl_tasks/__init__.py                 0      0   100%
parsl_tasks/cgcnn.py                   25     25     0%
parsl_tasks/compile_hull.py            63     63     0%
parsl_tasks/convex_hull.py            461    198    57%
parsl_tasks/dft_optimization.py        65     65     0%
parsl_tasks/ehull.py                  239    119    50%
parsl_tasks/gen_structures.py          72     72     0%
parsl_tasks/hull.py                    75     75     0%
parsl_tasks/select_structures.py      147     43    71%
tools/__init__.py                       0      0   100%
tools/config_labels.py                 29      0   100%
tools/config_manager.py               118     59    50%
tools/errors.py                         2      2     0%
tools/logging_config.py                39      2    95%
tools/post_processing.py               93     93     0%
workflows/__init__.py                   0      0   100%
workflows/vasp_assets/__init__.py       0      0   100%
workflows/vasp_based.py               101    101     0%
-------------------------------------------------------
TOTAL                                1909   1016    47%
43 passed, 5 warnings in 321.82s (0:05:21)
```

Our unit tests focus on pure-Python components (data prep, CGCNN wrappers, post-processing math). Some parts of the codebase are intentionally under-tested in CI because they require external systems:
- VASP
- An MP_API_KEY

I am currently working on an additional test for CGCNN.
