### 2025-10-08 ML-AMD/exa-amd#26 @PatrickJTaylor
Brilliant, thank you for the summary 😃

I understand that some of your codebase will be more challenging to unit test due to external dependencies, if you intend to leave those parts as is, then I think that's fair enough.

Please correct me if I'm wrong, but it looks to me like there is some untested code that does not have problematic external dependencies though? Just to give one example, the `cmd_compile_vasp_hull` function in `parsl_tasks/hull.py` does not require VASP to be run for testing, only that some example VASP outputs are available to be processed?

To be clear I'm not suggesting that you need to get to 100% coverage for all non-external-dependent code, I'm just trying to clarify if the majority of code *that you can test* is *being tested*.
