### 2025-10-20 #8879 @MoraruMaxim
Almost there! All changes have been finalized. We just need to complete the validation (PR review) and merge them into main. Thank you for checking!
