### 2025-10-21 ML-AMD/exa-amd#30 @MoraruMaxim
Great suggestion! I have opened #37 to address it.
Thank you for the insightful feedback!
