### 2025-10-22 ML-AMD/exa-amd#25 @shdejavu
Dear Reviewer,
Thank you for your time and for providing these insightful comments and constructive suggestions on our JOSS paper for exa-AMD. We appreciate your feedback, which will help us improve the clarity and impact of our manuscript.
We have addressed each of your suggestions below.
1: Is exa-AMD an alternative to automation frameworks like atomate2 and AiiDA, or is it a pre-configured materials discovery workflow?
This is an excellent point, and we recognize the need to clarify the scope and purpose of exa-AMD. Our intention is for exa-AMD to be a specialized, pre-configured materials discovery workflow, not a materials modeling/simulation automation framework.
In the field of computational materials science, several powerful workflow automation frameworks exist, such as AiiDA and atomate2. These tools provide a general-purpose engine for creating, managing, and executing complex sequences of calculations for different scientific tasks.
exa-AMD is not intended to be a direct alternative to these frameworks. Instead, it is a specialized, pre-configured scientific workflow to address the specific challenge of AI-assisted materials discovery. While general-purpose frameworks provide the building blocks to construct such a workflow, exa-AMD provides an integrated, out-of-the-box solution focused on a particular high-throughput screening methodology. Its modularity allows for customization within this context, but it is fundamentally an application-specific tool rather than a general-purpose workflow management system.
While exa-AMD is modular and configurable (allowing users to swap ML models, data sources, or post-processing tools), its core focus is to streamline high-throughput AI-assisted discovery rather than to replace or compete with existing automation frameworks. We have now made this distinction explicit in the revised manuscript:
“Unlike general-purpose workflow management systems such as AiiDA or atomate2, exa-AMD provides a domain-specific, end-to-end workflow tailored for AI-assisted materials discovery. Its modular design allows users to extend or replace components (e.g., machine learning models, data sources, or post-processing utilities), but its core goal is to deliver a pre-configured and scalable research pipeline optimized for materials discovery of multinary systems rather than serving as an materials modeling automation framework.”
We think it’s proper to add this paragraph in the summary section instead of adding a new “Related software” section.
2: Elaborate upon Figure 1
Thank you for pointing out the lack of detail in the caption for Figure 1. The Ce–Fe–In phase diagram shown in Figure 1 was generated using exa-AMD as a demonstration of its discovery capabilities. Ce₃FeIn emerged as a novel low-energy phase predicted to lie close to the convex hull (within a few meV/atom above hull), suggesting potential thermodynamic stability. This discovery exemplifies exa-AMD's ability to identify previously unreported intermetallic compounds in complex compositional spaces. We have revised the caption of Figure 1 for clarity:
“Figure 1: Example output from an exa-AMD workflow for the Ce-Fe-In chemical system. (Left) The predicted ternary convex hull, where each point represents a calculated crystal structure. The color indicates the energy above the hull (E_hull), with points on or very near the hull (in red, E_hull ≈ 0) predicted to be thermodynamically stable or metastable. (Right) The crystal structure of Ce₃FeIn, a new ternary compound identified by the workflow as a promising stable phase, located on the convex hull. This figure demonstrates the capability of exa-AMD to perform high-throughput screening and identify novel, potentially synthesizable materials.”
We also added one sentence in the "Workflow Overview" section in the main text:
“After the completion of VASP calculations, exa-AMD performs automated post-processing to extract and analyze key physical properties from the calculation outputs. This final stage computes the formation energies of each structure relative to reference elemental phases, which are then used for constructing the convex hull—the set of thermodynamically stable phases at zero temperature and pressure, where any compound above this hull is metastable or unstable [@ong2008li]. Structures with energy above the convex hull below a user-configurable threshold (typically 0.1 eV/atom) are identified as promising candidates and are automatically copied to a dedicated folder for further analysis. (Note that this criterion, based on thermodynamic stability relative to the convex hull, differs from the one used in the structure selection stage, where candidates are filtered based on their absolute formation energies, e.g., below a threshold of −0.2 eV/atom.) At the end of this stage, exa-AMD generates an updated phase diagram by plotting the convex hull. An example for the Ce-Fe-In system is shown in Figure 1, where exa-AMD identified several new potential compounds on the convex hull, including Ce₃FeIn.”
3: Clarify the caption of Figure 2
Thanks for the suggestions. Figure 2 represents the default/reference workflow template provided with exa-AMD. However, due to the modular architecture, users can modify, extend, or replace individual stages to suit their specific discovery objectives. And in response to #20 to clarify input/output parameters, we have revised the caption to make this clear:
“Figure 2: The core five-stage workflow implemented in exa-AMD. This represents a pre-configured pipeline for AI-assisted materials discovery. While this is the default workflow, the modular design of exa-AMD allows users to customize, replace, or extend individual stages to suit different research objectives. Inputs include target chemical elements and initial structure templates. Intermediate data includes ML-predicted formation energies and structural similarity metrics. Outputs comprise DFT-optimized structures, formation energies, energy above convex hull values, and updated phase diagrams.”

We have carefully reviewed all suggestions in comment #20 and response accordingly.

We believe these revisions directly address your concerns and will make the manuscript clearer for future readers. Thank you again for your valuable feedback.
Sincerely,
The exa-AMD Authors
