### 2025-10-23 #8879 @espottesmith
Hey @PatrickJTaylor, thank you for your update.

You raise a good question. If you'd like, I can go back to the editorial team with your question. But I'll give my personal interpretation: it's not reasonable to expect every reviewer to verify every result for every publication. We're all already drowning in peer review, and (again, this is my opinion) such a requirement would gum things up to a breaking point. What's important is that the results are _in principle_ verifiable and reproducible, such that, if there were any questions about their legitimacy, a diligent reviewer could pursue verification.

You say that you've tested the various parts of the software and that there's no reason in principle that the results in Figure 1 could not be reproduced. That seems reasonable to me.
