### 2025-11-09 #8879 @Kevin-Mattheus-Moerman
👉  No, this should be changed to `v0.1.0` as per releaes and archived version.

👉 there is one occurrence of `USA` please write this out as `United States of America`
