### 2025-11-10 #8879 @MoraruMaxim
Changed “USA” to “United States of America.”
Please note that the acronym “US” remains in “US DOE.”

Thank you!
