### 2025-11-17 #8879 @MoraruMaxim
Changed accordingly! Now both are `United States of America`. Thank you!
